create component files
