+++ frontend/src/components/admin/APIRequests.js
@@ -1,0 +1,11 @@
+import React, { useState, useEffect } from 'react';
+
+/**
+ * 
+ * @class APIRequests
+ */
+export default () => {
+    return (
+        <div></div>
+    )
+}
