+++ frontend/src/components/admin/AdminList.js
@@ -1,0 +1,11 @@
+import React, { useState, useEffect } from 'react';
+
+/**
+ * 
+ * @class AdminList
+ */
+export default () => {
+    return (
+        <div></div>
+    )
+}
