+++ frontend/src/components/admin/CloneList.js
@@ -1,0 +1,11 @@
+import React, { useState, useEffect } from 'react';
+
+/**
+ * 
+ * @class CloneList
+ */
+export default () => {
+    return (
+        <div></div>
+    )
+}
