+++ frontend/src/components/admin/HomepageEditor.js
@@ -1,0 +1,11 @@
+import React, { useState, useEffect } from 'react';
+
+/**
+ * 
+ * @class HomepageEditor
+ */
+export default () => {
+    return (
+        <div></div>
+    )
+}
